+++ src/main/java/ru/tn/profitcalculator/web/model/ProductGroup.java
@@ -16,6 +16,6 @@ public class ProductGroup {
     private Set<String> notes;
     private BigDecimal resultSum;
     private BigDecimal profitSum;
-    private Long optionProfitSum;
+    private BigDecimal optionProfitSum;
     private BigDecimal maxRate;
 }
